// src/openapi/client/models/GetChats200ResponseInner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My App
 * This is an awesome app!
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GetChats200ResponseInner
 */
export interface GetChats200ResponseInner {
    /**
     * 
     * @type {number}
     * @memberof GetChats200ResponseInner
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof GetChats200ResponseInner
     */
    user1?: number;
    /**
     * 
     * @type {number}
     * @memberof GetChats200ResponseInner
     */
    user2?: number;
}

/**
 * Check if a given object implements the GetChats200ResponseInner interface.
 */
export function instanceOfGetChats200ResponseInner(value: object): value is GetChats200ResponseInner {
    return true;
}

export function GetChats200ResponseInnerFromJSON(json: any): GetChats200ResponseInner {
    return GetChats200ResponseInnerFromJSONTyped(json, false);
}

export function GetChats200ResponseInnerFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetChats200ResponseInner {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'user1': json['user1'] == null ? undefined : json['user1'],
        'user2': json['user2'] == null ? undefined : json['user2'],
    };
}

export function GetChats200ResponseInnerToJSON(json: any): GetChats200ResponseInner {
    return GetChats200ResponseInnerToJSONTyped(json, false);
}

export function GetChats200ResponseInnerToJSONTyped(value?: GetChats200ResponseInner | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'user1': value['user1'],
        'user2': value['user2'],
    };
}