// src/openapi/client/models/GetMessages200ResponseInner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My App
 * This is an awesome app!
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GetMessages200ResponseInner
 */
export interface GetMessages200ResponseInner {
    /**
     * 
     * @type {number}
     * @memberof GetMessages200ResponseInner
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof GetMessages200ResponseInner
     */
    sender?: number;
    /**
     * 
     * @type {number}
     * @memberof GetMessages200ResponseInner
     */
    receiver?: number;
    /**
     * 
     * @type {string}
     * @memberof GetMessages200ResponseInner
     */
    content?: string;
    /**
     * 
     * @type {Date}
     * @memberof GetMessages200ResponseInner
     */
    createdAt?: Date;
}

/**
 * Check if a given object implements the GetMessages200ResponseInner interface.
 */
export function instanceOfGetMessages200ResponseInner(value: object): value is GetMessages200ResponseInner {
    return true;
}

export function GetMessages200ResponseInnerFromJSON(json: any): GetMessages200ResponseInner {
    return GetMessages200ResponseInnerFromJSONTyped(json, false);
}

export function GetMessages200ResponseInnerFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetMessages200ResponseInner {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'sender': json['sender'] == null ? undefined : json['sender'],
        'receiver': json['receiver'] == null ? undefined : json['receiver'],
        'content': json['content'] == null ? undefined : json['content'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
    };
}

export function GetMessages200ResponseInnerToJSON(json: any): GetMessages200ResponseInner {
    return GetMessages200ResponseInnerToJSONTyped(json, false);
}

export function GetMessages200ResponseInnerToJSONTyped(value?: GetMessages200ResponseInner | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'sender': value['sender'],
        'receiver': value['receiver'],
        'content': value['content'],
        'createdAt': value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
    };
}